+++ tests/testthat/test-reportPreview.R
@@ -62,7 +62,7 @@ test_that("Reactivity of reportPreview", {
     paste(collapse = ", ")
   
   str_expect <- "Vignettes, NEWS file, NEWS current, Report Bugs, Website, Maintainer, Source Control, Documentation, Bugs Closure Rate, License, First Version Release, Latest Version Release, Package Downloads, Reverse Dependencies, Monthly downloads trend"
-  
+
   expect_equal(maint_info, str_expect)
   
   app$stop()